feat: Enhance financial movements management with modal and API integration
- Updated the dashboard view to include a modal for creating financial movements.
- Implemented a new API endpoint for filtering financial movements.
- Added functionality to fetch the latest type and category based on description input.
- Improved the financial balances view with better currency formatting.
- Refactored financial movements form to support multiple wallets selection.
- Enhanced the user interface for financial categories with improved delete confirmation.
- Added a new table to display the last registered financial movements in the modal.
- Updated routes for financial movements to include new API endpoints.

+++ app/Http/Controllers/FinancialCategoryController.php
@@ -70,11 +70,12 @@ class FinancialCategoryController extends Controller
     public function destroy($id): RedirectResponse
     {
         try {
-            $category = FinancialCategory::findOrFail($id);
-            // if ($category->financials()->count() > 0) {
-            //     return Redirect::route('financial-categories.index')
-            //         ->with('error', __('Financial Category can not be deleted'));
-            // }
+            $category = FinancialCategory::find($id);
+
+            if ($category->financials()->exists()) {
+                 return Redirect::route('financial-categories.index')
+                     ->with('error', __('Financial Category can not be deleted'));
+            }
 
             $category->delete();
